更新网络配置文件创建逻辑

+++ src/spk/syno-network/src/udev/script/usb-netcard.sh
@@ -12,10 +12,14 @@ NAME=${2}
 createifcfg() {
   ETHX=${1}
   if [ -n "$(cat /usr/syno/etc/synoovs/ovs_reg.conf 2>/dev/null)" ]; then
-    echo -e "DEVICE=${ETHX}\nONBOOT=yes\nBOOTPROTO=dhcp\nIPV6INIT=dhcp\nIPV6_ACCEPT_RA=1\nBRIDGE=ovs_${ETHX}" >"/etc/sysconfig/network-scripts/ifcfg-${ETHX}"
-    echo -e "DEVICE=ovs_${ETHX}\nONBOOT=yes\nBOOTPROTO=dhcp\nIPV6INIT=dhcp\nIPV6_ACCEPT_RA=1\nPRIMARY=${ETHX}\nTYPE=OVS" >"/etc/sysconfig/network-scripts/ifcfg-ovs_${ETHX}"
     grep -q ${ETHX} /usr/syno/etc/synoovs/ovs_ignore.conf 2>/dev/null && sed -i "/${ETHX}/d" /usr/syno/etc/synoovs/ovs_ignore.conf
     grep -q ${ETHX} /usr/syno/etc/synoovs/ovs_interface.conf 2>/dev/null || echo ${ETHX} >>/usr/syno/etc/synoovs/ovs_interface.conf
+    if [ ! -f "/etc/sysconfig/network-scripts/ifcfg-${ETHX}" ]; then
+      echo -e "DEVICE=${ETHX}\nONBOOT=yes\nBOOTPROTO=dhcp\nIPV6INIT=dhcp\nIPV6_ACCEPT_RA=1\nBRIDGE=ovs_${ETHX}" >"/etc/sysconfig/network-scripts/ifcfg-${ETHX}"
+    fi
+    if [ ! -f "/etc/sysconfig/network-scripts/ifcfg-ovs_${ETHX}" ]; then
+      echo -e "DEVICE=ovs_${ETHX}\nONBOOT=yes\nBOOTPROTO=dhcp\nIPV6INIT=dhcp\nIPV6_ACCEPT_RA=1\nPRIMARY=${ETHX}\nTYPE=OVS" >"/etc/sysconfig/network-scripts/ifcfg-ovs_${ETHX}"
+    fi
     if [ ! "$(ovs-vsctl iface-to-br ${ETHX})" = "ovs_${ETHX}" ]; then
       ovs-vsctl br-exists ovs_${ETHX}
       [ $? -ne 2 ] && ovs-vsctl del-br ovs_${ETHX}
@@ -24,7 +28,9 @@ createifcfg() {
     fi
     ip link set ovs_${ETHX} up
   else
-    echo -e "DEVICE=${ETHX}\nONBOOT=yes\nBOOTPROTO=dhcp\nIPV6INIT=dhcp\nIPV6_ACCEPT_RA=1" >"/etc/sysconfig/network-scripts/ifcfg-${ETHX}"
+    if [ ! -f "/etc/sysconfig/network-scripts/ifcfg-${ETHX}" ]; then
+      echo -e "DEVICE=${ETHX}\nONBOOT=yes\nBOOTPROTO=dhcp\nIPV6INIT=dhcp\nIPV6_ACCEPT_RA=1" >"/etc/sysconfig/network-scripts/ifcfg-${ETHX}"
+    fi
   fi
 }
 
@@ -47,7 +53,7 @@ eth*)
     createifcfg "${ETHX}"
   else
     ip link set "${ETHX}" down
-    deleteifcfg "${ETHX}"
+    # deleteifcfg "${ETHX}"
   fi
   ovs-vsctl br-exists "ovs_${ETHX}" && /etc/rc.network "${ACTION}" "ovs_${ETHX}" || /etc/rc.network "${ACTION}" "${ETHX}"
   ;;
@@ -70,7 +76,7 @@ usb*)
     fi
   else
     ip link set "${ETHX}" down
-    deleteifcfg "${ETHX}"
+    # deleteifcfg "${ETHX}"
   fi
   ;;
 wlan*)
@@ -82,7 +88,7 @@ wlan*)
     createifcfg "${ETHX}"
   else
     ip link set "${ETHX}" down
-    deleteifcfg "${ETHX}"
+    # deleteifcfg "${ETHX}"
   fi
   ;;
 *)
